requirements.txt
langchain-openai
python-dotenv
langchain
streamlit